Increase handle size. Load resource defs
Former-commit-id: b83b964b78793a6dea5fb83223834b360a304476

+++ include/kb/freelist.h
@@ -8,19 +8,19 @@ extern "C" {
 #endif  
 
 typedef struct {
-  uint16_t*   handles;
-  uint16_t    count;
-  uint16_t    capacity;
+  uint32_t*   handles;
+  uint32_t    count;
+  uint32_t    capacity;
 } Freelist;
 
-KB_API void       kb_freelist_create        (Freelist* freelist, uint16_t capacity);
+KB_API void       kb_freelist_create        (Freelist* freelist, uint32_t capacity);
 KB_API void       kb_freelist_destroy       (Freelist* freelist);
 KB_API void       kb_freelist_reset         (Freelist* freelist);
-KB_API uint16_t   kb_freelist_take          (Freelist* freelist);
-KB_API void       kb_freelist_return        (Freelist* freelist, uint16_t handle);
-KB_API uint16_t   kb_freelist_count         (const Freelist* freelist);
-KB_API uint16_t*  kb_freelist_get_sparse    (Freelist* freelist);
-KB_API uint16_t*  kb_freelist_get_dense     (Freelist* freelist);
+KB_API uint32_t   kb_freelist_take          (Freelist* freelist);
+KB_API void       kb_freelist_return        (Freelist* freelist, uint32_t handle);
+KB_API uint32_t   kb_freelist_count         (const Freelist* freelist);
+KB_API uint32_t*  kb_freelist_get_sparse    (Freelist* freelist);
+KB_API uint32_t*  kb_freelist_get_dense     (Freelist* freelist);
 
 #ifdef __cplusplus
 }
